+++ src/components/Accordion/Accordion.tsx
@@ -14,12 +14,15 @@ import Icon from "../Icons/Icon";
 export type AccordionTypes = "default" | "warning" | "error";
 export interface AccordionDataProps {
   accordionType?: AccordionTypes;
+  ariaLabel?: string;
   label: string | JSX.Element;
   panel: string | React.ReactNode;
 }
 export interface AccordionProps {
   /** Array of data to display, and an optional accordionType */
   accordionData: AccordionDataProps[];
+  /** Description of aria-label. */
+  ariaLabel?: string;
   /** ID that other components can cross reference for accessibility purposes */
   id?: string;
   /** Whether the accordion is open by default only on its initial rendering */
@@ -62,6 +65,7 @@ const getIcon = (
  */
 const getElementsFromData = (
   data: AccordionDataProps[] = [],
+  ariaLabel: string,
   id: string,
   isAlwaysRendered: boolean = false,
   isDarkMode: boolean,
@@ -106,6 +110,16 @@ const getElementsFromData = (
         </AccordionPanel>
       );
 
+    const finalAriaLabel = content.ariaLabel ? content.ariaLabel : ariaLabel;
+
+    if (content.ariaLabel && ariaLabel) {
+      console.warn(
+        "NYPL Reservoir Accordion: And arialLabel value has been passed for the " +
+          "overall component and as part of the accordionData prop. Both can not " +
+          "be used, so the value in the accordionData prop will be used."
+      );
+    }
+
     return (
       <AccordionItem id={`${id}-item-${index}`} key={index}>
         {/* Get the current state to render the correct icon. */}
@@ -114,6 +128,7 @@ const getElementsFromData = (
           return (
             <>
               <AccordionButton
+                aria-label={finalAriaLabel}
                 id={`${id}-button-${index}`}
                 borderColor={
                   isDarkMode ? "dark.ui.border.default" : "ui.gray.medium"
@@ -185,6 +200,7 @@ export const Accordion = chakra(
   forwardRef<HTMLDivElement, AccordionProps>((props, ref?) => {
     const {
       accordionData,
+      ariaLabel,
       id,
       isDefaultOpen = false,
       isAlwaysRendered = false,
@@ -206,6 +222,7 @@ export const Accordion = chakra(
       >
         {getElementsFromData(
           accordionData,
+          ariaLabel,
           id,
           isAlwaysRendered,
           isDarkMode,